+++ musecpp/FieldBufferView.cpp
@@ -24,7 +24,7 @@ void FieldBufferView::set_prev_field(FieldBufferView *prev_field) {
     m_prev_field = prev_field;
 }
 
-void FieldBufferView::ProcessControlData(uint16_t const *control_data, std::pair<float, float> const &eq) {
+void FieldBufferView::ProcessControlData(float const *control_data, std::pair<float, float> const &eq) {
     m_control.emplace(ControlSignalDecoder(m_log, control_data, eq));
 }
